Add modules: more tests over modules
