+++ src/tests/modules/pal.c
@@ -1,0 +1,11 @@
+/* pal.h */
+
+#include "pal.h"
+#include "rev.h"
+#include <string.h>
+
+int palindrome(char *str) {
+  char reverseStr[100];
+  reverse(str, reverseStr);
+  return (strcmp(str, reverseStr) == 0);
+}
