+++ src/tests/modules/pal.h
@@ -1,0 +1,3 @@
+/* pal.h */
+
+int palindrome(char *str);
