+++ src/tests/modules/rev.h
@@ -1,0 +1,3 @@
+/* rev.h */
+
+void reverse(char *, char *);
